requirements.txt
transformers>=4.44.0
trl>=0.9.0
peft>=0.11.0
datasets>=2.20.0
bitsandbytes>=0.43.0
accelerate>=0.32.0
huggingface_hub>=0.23.0
PyYAML>=6.0.2
numpy>=1.26.0
torch>=2.3.0